requirements.txt
haystack-ai==2.1.2
opensearch-haystack
llama-cpp-python
llama-cpp-haystack
sentence-transformers==2.7.0